Fix positioning of extra admin tab buttons
