+++ src/ui/src/containers/admin/org-users.tsx
@@ -30,13 +30,15 @@ import {
 } from '@mui/material';
 import { Theme } from '@mui/material/styles';
 import { createStyles, makeStyles } from '@mui/styles';
-import { Link } from 'react-router-dom';
+import * as QueryString from 'query-string';
+import { Link, useHistory } from 'react-router-dom';
 
 import OrgContext from 'app/common/org-context';
 import { useSnackbar } from 'app/components';
 import { OAUTH_PROVIDER } from 'app/containers/constants';
 import { GQLOrgInfo, GQLUserInfo } from 'app/types/schema';
 
+import { InviteUserButton } from './invite-user-button';
 import {
   AdminTooltip, StyledTableCell, StyledTableHeaderCell, StyledRightTableCell,
 } from './utils';
@@ -57,6 +59,13 @@ const useStyles = makeStyles((theme: Theme) => createStyles({
   },
   approveButton: {
     width: theme.spacing(12),
+  },
+  inviteButtonWrapper: {
+    width: '100%',
+    padding: theme.spacing(2),
+    paddingTop: 0,
+    display: 'flex',
+    justifyContent: 'flex-end',
   },
   error: {
     padding: theme.spacing(1),
@@ -258,12 +267,24 @@ export const UsersTable = React.memo(() => {
   const users = data?.orgUsers || [];
   const classes = useStyles();
 
+  const history = useHistory();
+  const showInviteDialog = QueryString.parse(location.search).invite === 'true';
+  const onCloseInviteDialog = React.useCallback(() => {
+    if (showInviteDialog) {
+      // So that a page refresh after closing the dialog doesn't immediately open it again
+      history.replace({ search: '' });
+    }
+  }, [history, showInviteDialog]);
+
   if (error) {
     return <div className={classes.error}>{error.toString()}</div>;
   }
 
   return (
     <>
+      <div className={classes.inviteButtonWrapper}>
+        <InviteUserButton startOpen={showInviteDialog} onClose={onCloseInviteDialog} />
+      </div>
       {domainName && (
         <Alert icon={false} className={classes.managedDomainBanner} color='info' variant='outlined'>
           <p>
